3 - Backend/src/requirements.txt
Flask==2.0.3
pymysql==1.0.2
python-dotenv==0.19.2
